Add feature test for purchase return settlement with multiple line items at different prices
This test verifies that when returning products with multiple purchase detail lines at different prices, the settlement is calculated correctly based on the first line's price for the quantity returned. It ensures that the approval process accurately reflects the reduction in the purchase total, preventing overvaluation through flat averages.

+++ app/Livewire/PurchaseReturn/PurchaseReturnCreateForm.php
@@ -130,7 +130,7 @@ class PurchaseReturnCreateForm extends Component
                         'product_id' => $row['product_id'],
                         'product_name' => $row['product_name'],
                         'product_code' => $row['product_code'] ?? '',
-                        'quantity' => (int) $row['quantity'],
+                        'quantity' => (float) $row['quantity'],
                         'location_id' => $row['location_id'],
                         'unit_price' => (float) ($row['purchase_price'] ?? 0),
                         'price' => (float) ($row['purchase_price'] ?? 0),
